section 8: show order summary when clicking order
Adds a modal containing the order summary when the order button is
clicked. Not sure what a modal is, but is shows on top of the rest of
the page anyway and contains everything you ordered. It also comes with
a 50% transparent backdrop covering the rest of the page.

Also the modal has a cool slidey-slidey animation when it gets
enabled/disabled. Super neat-o.